# doc/src/main/java/com/example/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnlineShop API",
    "description": "java onlineshop",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/payment/request": {
      "post": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "requestPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoToPaymentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/product/{id}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/product/top/{type}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getTopProduct",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Popular",
                "Newest",
                "Cheap",
                "Expensive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseListProductDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payment/verify": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "verifyPayment",
        "parameters": [
          {
            "name": "Authority",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/product/category": {
      "get": {
        "tags": [
          "product-category-panel-controller"
        ],
        "operationId": "getAll_5",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListProductCategoryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/content": {
      "get": {
        "tags": [
          "content-panel-controller"
        ],
        "operationId": "getAll_8",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListContentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/color": {
      "get": {
        "tags": [
          "color-panel-controller"
        ],
        "operationId": "getAll_9",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "color",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListColorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/blog": {
      "get": {
        "tags": [
          "blog-panel-controller"
        ],
        "operationId": "getAll_10",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListBlogDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/content/{key}": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "operationId": "getByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseContentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Blog": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "operationId": "getAll_13",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseListBlogDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Blog/{id}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "operationId": "getById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseSingleBlogDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/product/{id}": {
      "delete": {
        "tags": [
          "product-panel-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/blog/{id}": {
      "delete": {
        "tags": [
          "blog-panel-controller"
        ],
        "operationId": "delete_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "APIResponseBoolean": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "boolean"
          }
        }
      },
...
      "SizeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ColorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "hex": {
            "type": "string"
          }
        }
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ProductCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "enable": {
            "type": "boolean"
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          },
          "exist": {
            "type": "boolean"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "addDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            },
            "uniqueItems": true
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeDto"
            },
            "uniqueItems": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategoryDto"
          }
        }
      },
      "APIResponseProductDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ProductDto"
          }
        }
      },
...
      "ContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "keyName": {
            "type": "string"
          },
          "valueContent": {
            "type": "string"
          }
        }
      },
      "APIResponseContentDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ContentDto"
          }
        }
      },
...
      "BlogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Published",
              "Draft"
            ]
          }
        }
      },
...
      "BasketItem": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "colorId": {
            "type": "integer",
            "format": "int64"
          },
          "sizeId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "GoToPaymentDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasketItem"
            }
          },
          "gateway": {
            "type": "string",
            "enum": [
              "Zarinpal",
              "CardToCard",
              "MellatBank",
              "TejaratBank",
              "PasargadBank"
            ]
          }
        }
      },
      "APIResponseString": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "APIResponseListProductDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          }
        }
      },
...
      "APIPanelResponseListProductCategoryDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCategoryDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "APIPanelResponseListContentDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "APIPanelResponseListColorDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "APIPanelResponseListBlogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "APIResponseListBlogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogDto"
            }
          }
        }
      },
      "APIResponseSingleBlogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SingleBlogDto"
          }
        }
      },
      "SingleBlogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}